backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Marib Aldoais <marib.aldoais@gmail.com>", "Luca Scalerandi <Scalerandi.Luca@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
logging_test = 'backend.logger:run'
validate_test = 'backend.validate_config:validate_config_structure'
process_test = 'backend.process_config:process_config'
add_phase_test = 'backend.process_config:add_phase_questions'
create_bucket_test = 'backend.process_config:create_file_storage'
add_test_users = 'backend.process_config:add_test_user'
run_script = 'backend.manage_applications:manage_users'

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.1"
logging = "^0.4.9.6"
loguru = "^0.7.3"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
python-dateutil = "^2.8.2"
supabase = "^2.0.0"
pytest = "^7.4.3"
pytz = "^2023.3.post1"
gotrue = "^2.1.0"
ruff = "0.11.2"


[tool.ruff]
line-length = 120
target-version = "py311"
lint.extend-select = ["I"]
lint.fixable = ["I"]
lint.isort.known-first-party = ["backend*"]
format.quote-style = 'single'


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
